backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
python-multipart==0.0.20
sse-starlette==2.2.1
imagehash==4.3.1
Pillow==11.1.0
